change layout for new batched version
